weight upload queue request by the number of multipart transfers
Previously, the number of concurrent uploads was limited only by the
number of concurrent "files". If any of the active transfers was of a
file over the multipart size limit, this would result in additional
active http sessions, which could result in the number of http sessions
being greater than the value of S3ND_UPLOAD_MAX_PARALLEL. This made it
difficult to control the maximum aggregate data rate.